# documentation/api/identity.openapi.yaml
openapi: 3.1.0
info:
  title: Zagrosi Identity API
  version: 0.1.0
  summary: Public identity, SSO, SCIM, and token-management API.
  description: >
    This specification documents the browser auth, OIDC, SAML, SCIM,
    session, personal access token, SCIM token, service token, and admin
    unlock surfaces shipped by the identity foundation.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
...
servers:
  - url: https://api.zagrosi.example
    description: Production API gateway
security: []
...
paths:
...
  /v1/orgs/{org_slug}/scim-tokens:
    get:
      tags: [Org SSO]
      operationId: listScimTokens
      summary: List org SCIM bearer tokens.
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      responses:
        "200":
          description: SCIM tokens.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScimTokenView"
    post:
      tags: [Org SSO]
      operationId: createScimToken
      summary: Mint a SCIM bearer. The raw `scim_` value is returned once.
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScimTokenRequest"
      responses:
        "201":
          description: SCIM token minted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedScimToken"
  /v1/orgs/{org_slug}/scim-tokens/{id}:
    get:
      tags: [Org SSO]
      operationId: getScimToken
      summary: Fetch SCIM token metadata.
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: SCIM token metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScimTokenView"
    delete:
      tags: [Org SSO]
      operationId: deleteScimToken
      summary: Revoke a SCIM bearer token.
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
        - $ref: "#/components/parameters/UuidId"
      responses:
        "204":
          description: SCIM token revoked.
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_sid
      description: Secure HttpOnly SameSite=Lax Path=/ session cookie.
    CsrfCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_csrf
      description: Double-submit CSRF cookie paired with `X-Zagrosi-CSRF`.
    OidcCsrfCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_oidc_csrf
      description: OIDC callback state cookie.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: "pat_<43>, scim_<43>, svc_<43>"
    ScimBearer:
      type: http
      scheme: bearer
      bearerFormat: scim_<43>
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OrgSlug:
      name: org_slug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9][a-z0-9-]{1,62}$"
...
  schemas:
...
    CreateScimTokenRequest:
      type: object
      required: [display_name]
      properties:
        display_name:
          type: string
    ScimTokenView:
      type: object
      required: [id, display_name, created_at]
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        created_at:
          type: string
          format: date-time
        revoked_at:
          type:
            - string
            - "null"
          format: date-time
    IssuedScimToken:
      allOf:
        - $ref: "#/components/schemas/ScimTokenView"
        - type: object
          required: [token]
          properties:
            token:
              $ref: "#/components/schemas/PrefixedToken"
...
    PrefixedToken:
      type: string
      pattern: "^(sid|pat|scim|svc|vrf|rst)_[A-Za-z0-9_-]{43}$"
      description: Token prefix is part of the SHA-256 hash input.
...